add CSW 3 client
