+++ owslib/catalogue/__init__.py
@@ -1,0 +1,7 @@
+# =============================================================================
+# Copyright (c) 2021 Tom Kralidis
+#
+# Authors : Tom Kralidis <tomkralidis@gmail.com>
+#
+# Contact email: tomkralidis@gmail.com
+# =============================================================================
