+++ owslib/fes2.py
@@ -12,7 +12,7 @@ API for OGC Filter Encoding (FE) constructs and metadata.
 
 Filter Encoding: http://www.opengeospatial.org/standards/filter
 
-Supports version 1.1.0 (04-095).
+Supports version 2.0.2 (09-026r2).
 """
 
 from owslib.etree import etree
@@ -23,19 +23,19 @@ from owslib.namespaces import Namespaces
 # default variables
 def get_namespaces():
     n = Namespaces()
-    ns = n.get_namespaces(["dif", "fes", "gml", "ogc", "xs", "xsi"])
-    ns[None] = n.get_namespace("ogc")
+    ns = n.get_namespaces(["dif", "fes", "gml", "ogc", "ows110", "xs", "xsi"])
+    ns[None] = n.get_namespace("fes")
     return ns
 
 
 namespaces = get_namespaces()
-schema = 'http://schemas.opengis.net/filter/1.1.0/filter.xsd'
-schema_location = '%s %s' % (namespaces['ogc'], schema)
+schema = 'http://schemas.opengis.net/filter/2.0/filterAll.xsd'
+schema_location = '%s %s' % (namespaces['fes'], schema)
 
 
 class FilterRequest(object):
     """ filter class """
-    def __init__(self, parent=None, version='1.1.0'):
+    def __init__(self, parent=None, version='2.0.0'):
         """
 
         filter Constructor
@@ -44,12 +44,12 @@ class FilterRequest(object):
         ----------
 
         - parent: parent etree.Element object (default is None)
-        - version: version (default is '1.1.0')
+        - version: version (default is '2.0.0')
 
         """
 
         self.version = version
-        self._root = etree.Element(util.nspath_eval('ogc:Filter', namespaces))
+        self._root = etree.Element(util.nspath_eval('fes:Filter', namespaces))
         if parent is not None:
             self._root.set(util.nspath_eval('xsi:schemaLocation', namespaces), schema_location)
 
@@ -65,7 +65,7 @@ class FilterRequest(object):
         - parent: the parent Element object.  If this is not, then generate a standalone request
         - qtype: type of resource to query (i.e. service, dataset)
         - keywords: list of keywords
-        - propertyname: the PropertyName to Filter against
+        - propertyname: the ValueReference to Filter against
         - bbox: the bounding box of the spatial query in the form [minx,miny,maxx,maxy]
         - identifier: the dc:identifier to query against with a PropertyIsEqualTo.  Ignores all other inputs.
 
@@ -181,9 +181,8 @@ class FilterRequest(object):
 
 
 class FilterCapabilities(object):
-    """ Abstraction for Filter_Capabilities """
+    """Abstraction for Filter_Capabilities 2.0"""
     def __init__(self, elem):
-        # Spatial_Capabilities
 
         if elem is None:
             self.spatial_operands = []
@@ -191,26 +190,33 @@ class FilterCapabilities(object):
             self.temporal_operators = []
             self.temporal_operands = []
             self.scalar_comparison_operators = []
+            self.conformance = {}
             return
 
-        self.spatial_operands = [f.text for f in elem.findall(util.nspath_eval(
-            'ogc:Spatial_Capabilities/ogc:GeometryOperands/ogc:GeometryOperand', namespaces))]
+        # Spatial_Capabilities
+        self.spatial_operands = [f.attrib.get('name') for f in elem.findall(util.nspath_eval(
+            'fes:Spatial_Capabilities/fes:GeometryOperands/fes:GeometryOperand', namespaces))]
         self.spatial_operators = []
         for f in elem.findall(util.nspath_eval(
-                'ogc:Spatial_Capabilities/ogc:SpatialOperators/ogc:SpatialOperator', namespaces)):
+                'fes:Spatial_Capabilities/fes:SpatialOperators/fes:SpatialOperator', namespaces)):
             self.spatial_operators.append(f.attrib['name'])
 
         # Temporal_Capabilities
-        self.temporal_operands = [f.text for f in elem.findall(util.nspath_eval(
-            'ogc:Temporal_Capabilities/ogc:TemporalOperands/ogc:TemporalOperand', namespaces))]
+        self.temporal_operands = [f.attrib.get('name') for f in elem.findall(util.nspath_eval(
+            'fes:Temporal_Capabilities/fes:TemporalOperands/fes:TemporalOperand', namespaces))]
         self.temporal_operators = []
         for f in elem.findall(util.nspath_eval(
-                'ogc:Temporal_Capabilities/ogc:TemporalOperators/ogc:TemporalOperator', namespaces)):
+                'fes:Temporal_Capabilities/fes:TemporalOperators/fes:TemporalOperator', namespaces)):
             self.temporal_operators.append(f.attrib['name'])
 
         # Scalar_Capabilities
         self.scalar_comparison_operators = [f.text for f in elem.findall(util.nspath_eval(
-            'ogc:Scalar_Capabilities/ogc:ComparisonOperators/ogc:ComparisonOperator', namespaces))]
+            'fes:Scalar_Capabilities/fes:ComparisonOperators/fes:ComparisonOperator', namespaces))]
+
+        # Conformance
+        self.conformance = {}
+        for f in elem.findall(util.nspath_eval('fes:Conformance/fes:Constraint', namespaces)):
+            self.conformance[f.attrib.get('name')] = f.find(util.nspath_eval('ows110:DefaultValue', namespaces)).text
 
 
 def setsortby(parent, propertyname, order='ASC'):
@@ -222,15 +228,15 @@ def setsortby(parent, propertyname, order='ASC'):
     ----------
 
     - parent: parent etree.Element object
-    - propertyname: the PropertyName
+    - propertyname: the ValueReference
     - order: the SortOrder (default is 'ASC')
 
     """
 
-    tmp = etree.SubElement(parent, util.nspath_eval('ogc:SortBy', namespaces))
-    tmp2 = etree.SubElement(tmp, util.nspath_eval('ogc:SortProperty', namespaces))
-    etree.SubElement(tmp2, util.nspath_eval('ogc:PropertyName', namespaces)).text = propertyname
-    etree.SubElement(tmp2, util.nspath_eval('ogc:SortOrder', namespaces)).text = order
+    tmp = etree.SubElement(parent, util.nspath_eval('fes:SortBy', namespaces))
+    tmp2 = etree.SubElement(tmp, util.nspath_eval('fes:SortProperty', namespaces))
+    etree.SubElement(tmp2, util.nspath_eval('fes:ValueReference', namespaces)).text = propertyname
+    etree.SubElement(tmp2, util.nspath_eval('fes:SortOrder', namespaces)).text = order
 
 
 class SortProperty(object):
@@ -241,9 +247,9 @@ class SortProperty(object):
             raise ValueError("SortOrder can only be 'ASC' or 'DESC'")
 
     def toXML(self):
-        node0 = etree.Element(util.nspath_eval("ogc:SortProperty", namespaces))
-        etree.SubElement(node0, util.nspath_eval('ogc:PropertyName', namespaces)).text = self.propertyname
-        etree.SubElement(node0, util.nspath_eval('ogc:SortOrder', namespaces)).text = self.order
+        node0 = etree.Element(util.nspath_eval("fes:SortProperty", namespaces))
+        etree.SubElement(node0, util.nspath_eval('fes:ValueReference', namespaces)).text = self.propertyname
+        etree.SubElement(node0, util.nspath_eval('fes:SortOrder', namespaces)).text = self.order
         return node0
 
 
@@ -252,7 +258,7 @@ class SortBy(object):
         self.properties = properties
 
     def toXML(self):
-        node0 = etree.Element(util.nspath_eval("ogc:SortBy", namespaces))
+        node0 = etree.Element(util.nspath_eval("fes:SortBy", namespaces))
         for prop in self.properties:
             node0.append(prop.toXML())
         return node0
@@ -275,45 +281,45 @@ class BinaryComparisonOpType(OgcExpression):
         node0 = etree.Element(util.nspath_eval(self.propertyoperator, namespaces))
         if not self.matchcase:
             node0.set('matchCase', 'false')
-        etree.SubElement(node0, util.nspath_eval('ogc:PropertyName', namespaces)).text = self.propertyname
-        etree.SubElement(node0, util.nspath_eval('ogc:Literal', namespaces)).text = self.literal
+        etree.SubElement(node0, util.nspath_eval('fes:ValueReference', namespaces)).text = self.propertyname
+        etree.SubElement(node0, util.nspath_eval('fes:Literal', namespaces)).text = self.literal
         return node0
 
 
 class PropertyIsEqualTo(BinaryComparisonOpType):
     """ PropertyIsEqualTo class"""
     def __init__(self, propertyname, literal, matchcase=True):
-        BinaryComparisonOpType.__init__(self, 'ogc:PropertyIsEqualTo', propertyname, literal, matchcase)
+        BinaryComparisonOpType.__init__(self, 'fes:PropertyIsEqualTo', propertyname, literal, matchcase)
 
 
 class PropertyIsNotEqualTo(BinaryComparisonOpType):
     """ PropertyIsNotEqualTo class """
     def __init__(self, propertyname, literal, matchcase=True):
-        BinaryComparisonOpType.__init__(self, 'ogc:PropertyIsNotEqualTo', propertyname, literal, matchcase)
+        BinaryComparisonOpType.__init__(self, 'fes:PropertyIsNotEqualTo', propertyname, literal, matchcase)
 
 
 class PropertyIsLessThan(BinaryComparisonOpType):
     """PropertyIsLessThan class"""
     def __init__(self, propertyname, literal, matchcase=True):
-        BinaryComparisonOpType.__init__(self, 'ogc:PropertyIsLessThan', propertyname, literal, matchcase)
+        BinaryComparisonOpType.__init__(self, 'fes:PropertyIsLessThan', propertyname, literal, matchcase)
 
 
 class PropertyIsGreaterThan(BinaryComparisonOpType):
     """PropertyIsGreaterThan class"""
     def __init__(self, propertyname, literal, matchcase=True):
-        BinaryComparisonOpType.__init__(self, 'ogc:PropertyIsGreaterThan', propertyname, literal, matchcase)
+        BinaryComparisonOpType.__init__(self, 'fes:PropertyIsGreaterThan', propertyname, literal, matchcase)
 
 
 class PropertyIsLessThanOrEqualTo(BinaryComparisonOpType):
     """PropertyIsLessThanOrEqualTo class"""
     def __init__(self, propertyname, literal, matchcase=True):
-        BinaryComparisonOpType.__init__(self, 'ogc:PropertyIsLessThanOrEqualTo', propertyname, literal, matchcase)
+        BinaryComparisonOpType.__init__(self, 'fes:PropertyIsLessThanOrEqualTo', propertyname, literal, matchcase)
 
 
 class PropertyIsGreaterThanOrEqualTo(BinaryComparisonOpType):
     """PropertyIsGreaterThanOrEqualTo class"""
     def __init__(self, propertyname, literal, matchcase=True):
-        BinaryComparisonOpType.__init__(self, 'ogc:PropertyIsGreaterThanOrEqualTo', propertyname, literal, matchcase)
+        BinaryComparisonOpType.__init__(self, 'fes:PropertyIsGreaterThanOrEqualTo', propertyname, literal, matchcase)
 
 
 class PropertyIsLike(OgcExpression):
@@ -327,14 +333,14 @@ class PropertyIsLike(OgcExpression):
         self.matchCase = matchCase
 
     def toXML(self):
-        node0 = etree.Element(util.nspath_eval('ogc:PropertyIsLike', namespaces))
+        node0 = etree.Element(util.nspath_eval('fes:PropertyIsLike', namespaces))
         node0.set('wildCard', self.wildCard)
         node0.set('singleChar', self.singleChar)
         node0.set('escapeChar', self.escapeChar)
         if not self.matchCase:
             node0.set('matchCase', 'false')
-        etree.SubElement(node0, util.nspath_eval('ogc:PropertyName', namespaces)).text = self.propertyname
-        etree.SubElement(node0, util.nspath_eval('ogc:Literal', namespaces)).text = self.literal
+        etree.SubElement(node0, util.nspath_eval('fes:ValueReference', namespaces)).text = self.propertyname
+        etree.SubElement(node0, util.nspath_eval('fes:Literal', namespaces)).text = self.literal
         return node0
 
 
@@ -344,8 +350,8 @@ class PropertyIsNull(OgcExpression):
         self.propertyname = propertyname
 
     def toXML(self):
-        node0 = etree.Element(util.nspath_eval('ogc:PropertyIsNull', namespaces))
-        etree.SubElement(node0, util.nspath_eval('ogc:PropertyName', namespaces)).text = self.propertyname
+        node0 = etree.Element(util.nspath_eval('fes:PropertyIsNull', namespaces))
+        etree.SubElement(node0, util.nspath_eval('fes:ValueReference', namespaces)).text = self.propertyname
         return node0
 
 
@@ -357,12 +363,12 @@ class PropertyIsBetween(OgcExpression):
         self.upper = upper
 
     def toXML(self):
-        node0 = etree.Element(util.nspath_eval('ogc:PropertyIsBetween', namespaces))
-        etree.SubElement(node0, util.nspath_eval('ogc:PropertyName', namespaces)).text = self.propertyname
-        node1 = etree.SubElement(node0, util.nspath_eval('ogc:LowerBoundary', namespaces))
-        etree.SubElement(node1, util.nspath_eval('ogc:Literal', namespaces)).text = '%s' % self.lower
-        node2 = etree.SubElement(node0, util.nspath_eval('ogc:UpperBoundary', namespaces))
-        etree.SubElement(node2, util.nspath_eval('ogc:Literal', namespaces)).text = '%s' % self.upper
+        node0 = etree.Element(util.nspath_eval('fes:PropertyIsBetween', namespaces))
+        etree.SubElement(node0, util.nspath_eval('fes:ValueReference', namespaces)).text = self.propertyname
+        node1 = etree.SubElement(node0, util.nspath_eval('fes:LowerBoundary', namespaces))
+        etree.SubElement(node1, util.nspath_eval('fes:Literal', namespaces)).text = '%s' % self.lower
+        node2 = etree.SubElement(node0, util.nspath_eval('fes:UpperBoundary', namespaces))
+        etree.SubElement(node2, util.nspath_eval('fes:Literal', namespaces)).text = '%s' % self.upper
         return node0
 
 
@@ -373,8 +379,8 @@ class BBox(OgcExpression):
         self.crs = crs
 
     def toXML(self):
-        tmp = etree.Element(util.nspath_eval('ogc:BBOX', namespaces))
-        etree.SubElement(tmp, util.nspath_eval('ogc:PropertyName', namespaces)).text = 'ows:BoundingBox'
+        tmp = etree.Element(util.nspath_eval('fes:BBOX', namespaces))
+        etree.SubElement(tmp, util.nspath_eval('fes:ValueReference', namespaces)).text = 'ows:BoundingBox'
         tmp2 = etree.SubElement(tmp, util.nspath_eval('gml:Envelope', namespaces))
         if self.crs is not None:
             tmp2.set('srsName', self.crs)
@@ -405,23 +411,23 @@ class BinaryLogicOpType(OgcExpression):
 
 class And(BinaryLogicOpType):
     def __init__(self, operations):
-        super(And, self).__init__('ogc:And', operations)
+        super(And, self).__init__('fes:And', operations)
 
 
 class Or(BinaryLogicOpType):
     def __init__(self, operations):
-        super(Or, self).__init__('ogc:Or', operations)
+        super(Or, self).__init__('fes:Or', operations)
 
 
 # UNARY
 class UnaryLogicOpType(OgcExpression):
     """ Unary Operator: Not """
-    def __init__(self, urary_operator, operations):
-        self.urary_operator = urary_operator
+    def __init__(self, unary_operator, operations):
+        self.unary_operator = unary_operator
         self.operations = operations
 
     def toXML(self):
-        node0 = etree.Element(util.nspath_eval(self.urary_operator, namespaces))
+        node0 = etree.Element(util.nspath_eval(self.unary_operator, namespaces))
         for op in self.operations:
             node0.append(op.toXML())
         return node0
@@ -429,4 +435,4 @@ class UnaryLogicOpType(OgcExpression):
 
 class Not(UnaryLogicOpType):
     def __init__(self, operations):
-        super(Not, self).__init__('ogc:Not', operations)
+        super(Not, self).__init__('fes:Not', operations)
