+++ owslib/util.py
@@ -407,7 +407,7 @@ def http_post(url=None, request=None, lang='en-US', timeout=10, username=None, p
     headers = {
         'User-Agent': 'OWSLib (https://geopython.github.io/OWSLib)',
         'Content-type': 'text/xml',
-        'Accept': 'text/xml',
+        'Accept': 'text/xml,application/xml',
         'Accept-Language': lang,
         'Accept-Encoding': 'gzip,deflate',
         'Host': u.netloc,
